requirements.txt
APScheduler==3.6.3
Babel==2.8.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
cloudinary==1.18.1
cryptography==2.8
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.6
django-apscheduler==0.3.0
django-background-task==0.1.8
django-background-tasks==1.2.0
django-common-helpers==0.9.2
django-compat==1.0.15
django-cors-headers==3.1.1
django-heroku==0.3.1
django-icons==1.0.0
django-phonenumber-field==4.0.0
django-storages==1.8
django-twilio==0.12.1
django-uuidfield==0.5.0
djangorestframework==3.10.3
dropbox==9.4.0
fcm-django==0.3.2
gcloud==0.17.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
googlemaps==3.1.3
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
imageio==2.6.1
jws==0.1.3
mock==3.0.5
numpy==1.17.3
oauth2client==3.0.0
onesignal-sdk==1.1.0
phonenumbers==8.11.2
Pillow==6.2.0
proglog==0.1.9
protobuf==3.10.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.4.3
pyfcm==1.4.7
PyJWT==1.7.1
pyOpenSSL==19.0.0
Pyrebase==3.0.27
python-http-client==3.2.1
python-jwt==2.0.1
python-magic==0.4.15
pytz==2019.2
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
sendgrid==6.1.0
six==1.13.0
sqlparse==0.3.0
tqdm==4.37.0
twilio==6.35.3
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
